Start work on saving/restoring database from a file
This will be done using a binary file.  This will allow the file to stay small.  Each entry in the file will consist of a integer for the id, an int for the number of key/value pairs there are then the keys and values as strings with an int first specifying the length.  This simplifies the reading process and allows for the correct amount of memory to be allocated.