Removed dbus service and moved the MediaObject creation into the library.
The MediaObject is created by the library.  If another application already has ownership of the object, it creates a proxy to that object.  It will then watch the bus for name changes.  When a the org.gnome.GMediaDB.'type' name owner changes it reacts and either creates a new proxy or accesses it's MediaObject directly.