remove get_tags function and change behavior of get_entry methods when tags is null.
The functions used to get entries now behave differently when the tags option is
null.  Now they return an array of char* that include both the tag and value.
It is now a null terminated list like key1, val1, key2, val2, ..., NULL.  This
is used when you want to get all of the tags for a particular entry.